+++ ryzen-smu-cli/Hardware/IRyzenController.cs
@@ -14,6 +14,10 @@ internal interface IRyzenController : IDisposable
 
     bool CanWritePboOffsets { get; }
 
+    bool CanReadFMax { get; }
+
+    bool CanWriteFMax { get; }
+
     OperationResult<int> GetPboOffset(CoreAddress core);
 
     OperationResult SetPboOffset(CoreAddress core, int offset);
@@ -21,6 +25,10 @@ internal interface IRyzenController : IDisposable
     OperationResult<float> GetPboScalar();
 
     OperationResult SetPboScalar(int scalar);
+
+    OperationResult<uint> GetFMax();
+
+    OperationResult SetFMax(uint megahertz);
 
     DowncoreOperationResult SetDisabledCores(IReadOnlySet<int> physicalCoreIndices);
 }